// src/enums/actions-enum.ts
export enum Actions {
  SET_MESSAGES = 'SET_MESSAGES',
  LOAD_DB_MESSAGES = 'LOAD_DB_MESSAGES',
  SET_INPUT = 'SET_INPUT',
  SET_USER_ID = 'SET_USER_ID',
  SET_LOADING = 'SET_LOADING',
  SET_CHAT_SETTINGS_OPEN = 'SET_CHAT_SETTINGS_OPEN',
  SET_CHAT_SETTINGS_USER_MESSAGE_BG = 'SET_CHAT_SETTINGS_USER_MESSAGE_BG',
  SET_CHAT_SETTINGS_BOT_MESSAGE_BG = 'SET_CHAT_SETTINGS_BOT_MESSAGE_BG',
  SET_CHAT_SETTINGS_USER_MESSAGE_COLOR = 'SET_CHAT_SETTINGS_USER_MESSAGE_COLOR',
  SET_CHAT_SETTINGS_BOT_MESSAGE_COLOR = 'SET_CHAT_SETTINGS_BOT_MESSAGE_COLOR',
  SET_CHAT_SETTINGS_BOT_NAME = 'SET_CHAT_SETTINGS_BOT_NAME',
  SET_IS_FIRST_MESSAGE = 'SET_IS_FIRST_MESSAGE',
}